+++ services/triton/service.go
@@ -7,7 +7,7 @@ import (
 	"time"
 
 	"github.com/influxdata/kapacitor/services/scraper"
-	"github.com/prometheus/prometheus/config"
+	"github.com/prometheus/prometheus/discovery/targetgroup"
 	ptriton "github.com/prometheus/prometheus/discovery/triton"
 )
 
@@ -133,7 +133,7 @@ func (s *Service) Test(options interface{}) error {
 	}
 
 	ctx, cancel := context.WithCancel(context.Background())
-	updates := make(chan []*config.TargetGroup)
+	updates := make(chan []*targetgroup.Group)
 	go discoverer.Run(ctx, updates)
 
 	select {